refactor(api): getPartnershipList을 review.js에서 partnership.js로 이동
